init: set up project
