+++ tsconfig.json
@@ -1,0 +1,14 @@
+{
+	"compilerOptions": {
+	  "target": "ES2020",
+	  "module": "Node16",
+	  "moduleResolution": "node16",
+	  "outDir": "./dist",
+	  "rootDir": "./src",
+	  "esModuleInterop": true,
+	  "strict": true,
+	  "skipLibCheck": true
+	},
+	"include": ["src/**/*.ts"],
+	"exclude": ["node_modules", "dist"]
+  }
